Updated openstack/openstack Project: openstack/nova  40c675b3d883cc730385faca513b3025a83a2e9b
Stop explicitly running test discovery for py34

This commit removes the direct call to discovery in the tox py34 job.
This was added as a temporary step to enable importing of everything
and is now include by standard in the use of ostestr. There isn't any
reason to keep it around anymore.

Change-Id: I117fbc2108b285aa148cb447c1c01145c43031dd