Updated openstack/openstack Project: openstack/nova  f07149f0f01f1c7962c4742264f6f6bdfeb20899
VMware: add in folder support on VC

In an effort to provide more information to VC admins we create a
a folder that will contain the OpenStack information. VM's will be
put in the following folder:
    Openstack/Project (<project-id>)/Instances

Until this patch the VM's were put in the root folder.

NOTE: the code is backward compatible. All newly created instances
will be placed in the aforementioned folder. Existing instances will
continue to work as usual.

This is the first part of the blueprint vmware-better-display-names

Change-Id: I6c653cf580e844f121a6566c4f2cbe320531f4e4