Updated openstack/openstack Project: openstack-infra/system-config  2bca224b7c83862b9e2b5ed31eaad75510fc8953
Redirect puppet cron output to a log file

The _ansible_ output (mostly consisting of puppet output) is now sent
to the puppet_run_all log file, but there is still quite a bit of
output from ancilary parts of the cron job that just gets sent in
email to infra-root every few minutes.  Send it to a log file.

Change-Id: I1f382b2d376cbfc2eff430c897bd3737fd585373