Updated openstack/openstack Project: openstack/ironic  b48289e8a230c593ded6a2aecfbc8190f790b784
Remove release differences from flavor creation docs

The documentation in the master should reflect to the current version of
the code, therefore information from previous releases should be
removed. This patch is simplifying the flavor creation section of the
documentation because it reflected to many different versions of Ironic,
since Icehouse.

Change-Id: I5d3d54783dcaf1828cf8e09efcf2239566e5a916
Partial-Bug: #1523962