Updated openstack/openstack Project: openstack-dev/devstack  88509e796c650d882ee26df0eed1f5bbb7f69de9
Fix restoring a backup volume to a larger volume

We cannot restore a backup to a larger volume on ceph because it
fails with status "error_restoring". This patch adds read/write
permissions to volumes pool for backup user. We need such permissions
to change volume size during restoring backup when the backup is
smaller than a volume.

Change-Id: I794c1126bcee4e07baf5a9dcfee779fd61da5636
Closes-Bug: 1519749