Updated openstack/openstack Project: openstack-dev/devstack  f3c722fb0a4b18744b84fe564d60f3f076e3679e
Make image_list.sh independent of host IP address

We do not need the HOST_IP to be detected in order to be able to list
our images. So just set that to some dummy value before sourcing
functions.

This will allow tools like disk-image-builder to work regardless of
whether get_default_host_ip succeeds or not.

Change-Id: I9c22d2066e34309e70e56076e3d17c5db6ecee06