Updated openstack/openstack Project: openstack/cinder  ca273277d45e5708fbc2aaf0dd4d0aa447b4c4f4
Optimize 3PAR array ID retrieval

Currently, we are making a call to the 3PAR array every time the
common driver is initialized, which is inefficient considering the
call is already made in get_volume_stats. Instead of making the call
every time, we can pass in the driver stats to common's do_setup and
use the array ID from there.

If the stats are not yet available, we will ask the array directly for
the ID.

Change-Id: Iff065cd9debfc899b1d88475c63c7b2cb1e5e1b6