Updated openstack/openstack Project: openstack/cinder  cb14dbeaf89fe3235a197e53954c751c26f67973
Change minimum 3PAR API version for replication

Currently, the primary and all secondary 3PAR arrays must have an API
version of 1.4.2 or higher and all be on the same version. This is
changed to require version 1.5.0 or higher and does not force all
arrays to be on the same version.

Change-Id: Ie6cc19f3bc4939e1ad15b64140a21bab1bc27fc7