Updated openstack/openstack Project: openstack/trove  f58b12e776393ea0164191579797ec61168b846c
Fix race condition in replication backup delete

In the int-tests, the replication implementation verifies that the
backup used to create the replica is removed after the slave goes
active. This test runs right after the slave create, so on occasion
the backup still exists when the test runs, since the taskmanager
doesn't delete it until it also receives word that the instance is
up and running.

The test was modified to wait until the backup is gone.

The replication scenario test was also modified to include this
check at the end of its run.

Closes-Bug: #1518477
Change-Id: If815f496d6bf9111dd1f04bf02f61207a2317b2b