Updated openstack/openstack Project: openstack-infra/project-config  dcf6181665254a1b164dfaacaa272c31742ccb37
Constraint based experimental jobs for neutron-vpnaas

Adding additional experimental jobs for neutron-vpnaas repo for
running with constraints.  This is the first step, in adding
constraint based support to neutron-vpnaas.  Later commits will
update tox.ini in the repo to support the new jobs, make the
jobs non-voting and runnable for releases after Kilo, and then
make the jobs voting.

Removed jobs that don't currently work:
 - gate-grenade-dsvm-neutron-vpnaas
 - gate-tempest-dsvm-neutron-pg-full

This will be the template for what we can do for other advanced
services too.

Change-Id: I3a2c3c81533ce6145cd16c2a08a4a2af0dfb0d07
Partial-Bug: #1522503