Updated openstack/openstack Project: openstack/ironic-specs  39176e4392b7d06066ab9aa4b39a2b3c7155fd5e
Update template to include openstackclient

In addition to the ironic CLI, we have added the openstackclient CLI
(http://specs.openstack.org/openstack/ironic-specs/specs/approved/ironicclient-osc-plugin.html).

This adds two subsections in the CLI section of the template, to
describe changes to these two CLIs. Note that because they are
subsections, our unit tests will not enforce the presence of these
two sections.

Change-Id: Ie07a41c9815fbf7d2bd14ae9830a92b7db2d01f4