Updated openstack/openstack Project: openstack-infra/tripleo-ci  6eb3a7ac56998a0dd623304cfc1fe2029c912b1f
Don't pass duplicate projects to tripleo.sh --delorean-build

Currently we update the list of projects to build for every repo ref
but we actually only need to add the project once for delorean to
build it.  Delorean actually skips any already-built projects, but
this cleans up the logs when building long dependency chains.

Change-Id: Ie9e6bd828e54e0dd9a6a737e016aa429b87a82a5