Updated openstack/openstack Project: openstack-infra/project-config  d0e14a39103917dde7c4d9df8fd1ad1a21a83976
Add Manila with CephFSNative driver CI job

Add Jenkins job for Manila with CephFSNative as its storage backend
driver.

Set the job as experimental and non-voting for the Manila project
as the CephFSNative driver is still WIP.

Set the job as non-voting for the devstack-plugin-ceph project until
the job is proven to be stable.

Change-Id: Ia3221b735468bd7ab6b2d4dccba75cae76e09dd4