Updated openstack/openstack Project: openstack-infra/system-config  df2fafecd8e9d1fe8a98008322755b928d5e022f
Add use_hostnames to ansible clouds.yaml

This is the clouds.yaml file that gets used by the ansible openstack
inventory system.  Setting use_hostnames to true indicates that
generally, our hostnames are expected to be unique, so ansible can
add them to the inventory based on their hostnames rather than their
uuids.  In the case that we have more than one host with the same
hostname, ansible will still favor uuids.

This should cause our logs to be more readable as they will include
hostnames for most of the ansible actions.

Change-Id: Ie4c24317429a8f0ac59729a2065dfae7fc6679d9