Updated openstack/openstack Project: openstack/cinder  34ed2611d05e2ac4c8ebc13c54b494140172f2cd
Enable consisgroups in SolidFire driver

This patch enables Consistency Group support in the SolidFire driver.
SolidFire group snapshots are what is used to provide consistent
snapshots on the backend.

A variety of helpers were used to wrap the API requests, meanwhile
other helpers were utilized to jump through the necessary hoops to
adopt the SolidFire API/group snapshots for use in consistency groups.
Because the SolidFire array does not create logical groupings of
volumes as consistency groups, there are a lot of no-ops used to
satisfy the necessary Cinder functions as specified in driver.py.
In-line comments attempt to clarify this. In short, a lot of searching
is needed to create consistency groups on the fly per request.

Change-Id: I82a8b2b6f2db4fca0c1cac892e323eae59c6b98c