Updated openstack/openstack Project: openstack/docs-specs  acb063126514f327e81ee7a29925abca1c467434
Adjusting a sentence in the User Guide Specification

The original Specification suggested a proposed change of
merging the Cloud Admin Guide into the Administrator User Guide.
I wrote this suggestion with the doc titles in the wrong order.
This patch corrects the order. The intended goal is merging
Admin User Guide content into the Cloud Admin Guide.

Change-Id: I034a9f1e0900f491c76aa989ce0ea70ae4d3f564