Updated openstack/openstack Project: openstack/heat  c5549259ba7633f4ec2faaa204dd7d4039668a9f
Fix race condition in tests with swift key

In integration we publish templates in Swift using temporary URLs.
Unfortunately there is a race condition in the way we setup the key, as
we use the key global to the account. This switches to a container per
test and a key per container to work around that.

Change-Id: Ia19967696811847e5b3ba07df68848a1708c2316
Closes-Bug: #1491773