Updated openstack/openstack Project: openstack/tempest  0dd1c7bb5d60e0fe2cbf7cf4960ff53b71179146
remove log_console_of_servers from get_remote_client

This is never used by any caller of this function, and just causes
confusion. It also explains where console logs in scenario tests are
not as useful as they could be because we never can scope the console
logs to the guest that failed.

Change-Id: I9ff86a5b1c8c31ad4bb9b86fcda59cfeae954ef2