Updated openstack/openstack Project: openstack/tempest-lib  3f49a48090a05448d17e64349eeee8c899f6566b
Migrated quotas_client and security_groups_client

This migrates the above files from tempest.
This includes tempest commits:

 * quotas_client.py         : I8087a7bf5e130778df9629065a8bfda8018fdbe0
 * security_groups_client.py: I88a43f44346671de5dfb759d07c4086cffa8d4d9

to see the commit history for these files refer to the above Change-Ids
in the tempest repository.

NOTE: reset_quotas() of quotas_client is different from the method
      naming rule, but is is the same as the api-site. So it is fine
      now.

Partially implements blueprint migrate-service-clients-to-tempest-lib

Change-Id: I559c7ec29d6cc85060d59b8b7714a669a277f217