Updated openstack/openstack Project: openstack/neutron  e534c88452fe0616a472eedc343cdd6b7f96ed3c
fix call which is only specific to enhanced_rpc

security_group_updated method is not implemented in the default
FirewallDriver class. But SecurityGroupAgentRpc class assumes the method
is implemented by the underlying Firewall driver.

This patch adds a check to ensure that only the drivers which leverages
enhanced RPC will make the call.

Related-Bug: #1493104
Change-Id: I0ee44a5ce69368c7731adfe9953adc0b2c5ff2d2