Updated openstack/openstack Project: openstack/keystone  4d62e5d40027c60eafa86ef42d5618e9e6e10107
Fix users in group and groups for user exact filters

List users in a group by name and list groups for user by name
throws HTTP 500 error in case of exact filters because filter_by
query used in case of exact filter is not able to filter using
'name' attribute.
This patch fixes this issue by replacing filter_by with filter query
in case of exact_filter similar to the inexact_filter query.

Closes-Bug: 1521772
Change-Id: Ib44513540370c0701a439f11a49e0dc338601fae