Updated openstack/openstack Project: openstack-dev/devstack  828936c6ac8045fbb985ed2e89cdaa9a7c01241a
Use keystone[ldap] for ldap packages

Keystone now provides an "ldap" in extras to install its ldap
dependencies so devstack doesn't have to track the python
dependencies itself.

Installation of the extras is done in an extra install line. This is
slightly redundant, however this pattern works much better from an
install stand point as it supports an arbitrarily large number of
extras.

Partial-Bug: 1479962
Change-Id: If0f0ff48f3d6b3c414f2d6fcd747ecf45a397658