Updated openstack/openstack Project: openstack/neutron  0d77401ceb6bbcf9035e53a0def43053053b0036
Create a routing table manager

The routing table manager maps address scope ids to routing tables.
It uses the rt_tables file specific to each namespace to maintain the
mapping so that id can simply be used as the table name when running
iproute2 commands.  This will be useful when debugging.

Change-Id: Icd5e98c82a070045a50e0c5d3762906b7e159d3d
Partially-Implements:  blueprint address-scopes