Updated openstack/openstack Project: openstack/nova  ce34000c85f0f2f481b894b8ce57acef7612770b
libvirt: update to min required version to 0.10.2

As declared during Liberty, the new min required libvirt
version is now 0.10.2.

NB, NEXT_MIN_LIBVIRT_VERSION now matches the current
MIN_LIBVIRT_VERSION. This constant is left in the
code since it is expected we'll repeatedly use it
in future to warn about subsequent version bumps.

The proposed min version for Nxxx is 1.2.1 which
allows Debian Wheezey Backports as the oldest
supported distro.

Change-Id: If7dcfdae296fab44ae44d670eea4234ad9381d6d