Updated openstack/openstack Project: openstack/nova  0f152d11102732fdf4c1906f861537201411d3d1
scheduler: add soft-(anti-)affinity weighers

ServerGroupSoftAffinityWeigher implements the soft-affinity
policy for server groups as it orders the hosts by the number
of instances running on them from the same group in decreasing
order.

ServerGroupSoftAntiAffinityWeigher implements the
soft-anti-affinity policy for server groups as it orders the
hosts by the number of instances running on them from the same
group in increasing order.

Both weigher assumes the the RequestSpec object refers to a valid
InstanceGroup object which has a populated members field. This
will be provided by the subsequent patch.

New configuration options added:
 * soft_affinity_weight_multiplier
 * soft_anti_affinity_multiplier

Add docs about the new weigher classes and change the function
name 'weigh_object' to '_weigh_object' in the Weights section
to match the code.

DocImpact

Co-Authored-By: Balazs Gibizer <balazs.gibizer@ericsson.com>

Implements: blueprint soft-affinity-for-server-group
Change-Id: I3f156d5e5df4d9642bb4b0ffac30a6288459ce61