Updated openstack/openstack Project: openstack-infra/project-config  e0b6fa5c182fc7ce27014c76da0501045f324658
Validate grafyaml based on master

Rather than installing grafyaml from pypi, install it with
zuul-cloner so that it may be validated with the contents of
master (or even changes in review).  We run grafyaml continuously
deployed from source rather than installing from pypi, so this
more closely matches what we actually want to validate.

Change-Id: I26663e5c70883b910fced6c32ae364734542b5a5