Updated openstack/openstack Project: openstack/trove-integration  900796e947f9d4a24b801844d482c375e9d1aa5e
images wont build because percona-xtrabackup-22 not found

This is using the ubuntu apt repo for the percona xtrabackup.
Making this change should fix it for now until we change to use
the percona repo for clustering.

Also the fedora repo does not have the percona-xtrabackup-22
package either but only the percona-xtrabackup package.

Closes-Bug: #1514933
Change-Id: I05aa4ee732b8fe8363f2d10a774ad4a2e11eca9d