Updated openstack/openstack Project: openstack-infra/project-config  627b9e48f2233f7ca18e860f2c2e9ec008b2a7ac
Add solum-pythonclient to projects

We are missing solum-pythonclient in $PROJECTS variable
and our gate is currently failing.  This will unblock
it and allow us to submit a number of otherwise good patches.

Also added SOLUM_PROJ_DIR to $DEVSTACK_LOCAL_CONFIG in order
to utilize it in our devstack-gate setup for sample config
generation.

Added KEEP_LOCALRC to use already existing localrc file

Change-Id: Id5e6b214afa693d6d5a02bc33286ddacad6289d2