Updated openstack/openstack Project: openstack-infra/project-config  221cb66047ba6302af7d914ff5821ec2be435e5c
Fedora 21 is EOL stop using it

Fedora 21 reached end of life on December 1, 2015. We should not be
testing on end of life distro releases so remove tests that run on
Fedora 21. A follow up change should completely remove the configuration
from nodepool once we have deleted all images existing slaves.

Change-Id: Iab7e4676d534e59848835a1874843705ae301955