Updated openstack/openstack Project: openstack/python-troveclient  3bace1b33cca84d3835b5419d643c4a8a24c4268
Put py34 first in the env order of tox

To solve the problem of "db type could
not be determined" on py34 we have to run first the py34 env to, then,
run py27. This patch puts py34 first on the tox.ini list of envs to
avoid this problem to happen.

Change-Id: I90be329de961a0a63607d04b5c7cc88d033f05ec
Closes-bug: #1489059