Updated openstack/openstack Project: openstack/trove  19af8761a8e5807401f8faefe0dd69ff77c7f542
Trivial: Remove meaningless default value in __get__()

In function object.__get__(self, instance, owner), owner is always
the owner class, while instance is the instance that the attribute
was accessed through, or None when the attribute is accessed through
the owner. Setting owner default value makes no sense.
refer to:
http://python-reference.readthedocs.org/en/latest/docs/dunderdsc/get.html#get

Change-Id: I5b84c77a52e5b26cb76d70ac738e1caef6e74c09
Closes-Bug: #1531395