Updated openstack/openstack Project: openstack/nova  2a5c85401383c1c9e1286b957760697c48a1a6e4
Remove null AZ tests from API tests

The AggregateAPI will never return an aggregate object with a null
availability zone (which is stored as part of the metadata field). Some
tests are stubbing out the AggregateAPI to return a fake aggregate
object with a null availability zone, which is an invalid state. The
Aggregate object will also always reject setting the availability zone
in the metadata to None by raising a ValueError.

Because of this, these tests are invalid and are removed.

Change-Id: I0cf469875b9d3f3b7a5e74aa169964793dc7d6ba