Updated openstack/openstack Project: openstack/python-glanceclient  dfff3af389cc1404a4e8860b717b9ede9839069b
Remove py26 support

As of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I50eff4ea33358791cc4f139b3b369489c38f7e5c
Closes-Bug: 1519510