Updated openstack/openstack Project: openstack/oslo.utils  af49730dabea64988ceabc06e1a171bd9ea99d03
Add "configdrive" to the list of keys used by mask_password()

The node representation in Ironic can contain a key called "configdrive"
which the value can be either a base64 string or a Swift temporary URL
(which is a secret). We need to mask that key when logging the node
representation either in Ironic or Nova driver so reusing the oslo's
mask_password() makes sense.

Change-Id: I10f9f37ff635f9a2dddfd7f7aaaac7d4fe6244b9