Updated openstack/openstack Project: openstack/nova  7bd0bc7b4c11cb2fb40d29c2da94a59f76fd0150
XenAPI: Cope with more Cinder backends

Some Cinder backends (e.g. storwize 7000) expose multiple LUNs for a single
SR connection from the XenServer host.  Since the SR UUID was based on the
'volume' ID, which is unique per volume, multiple SRs would be connecting
to the same array when multiple volumes are exposed.

This fixes the issue by making sure that we only connect to the array once
for each (host,port,IQN) tuple, and we can re-use the SR for as many LUNs
are presented.

Change-Id: I6a1eb22fcfde8e3f5c069be54f5e7c91b0f2fec1
Closes-Bug: 1515672