Updated openstack/openstack Project: openstack/neutron  5659fe1e829a8fecef664bc5e1c9df50a8504ef0
Fix uuid passing in disable_isolated_metadata_proxy

In DhcpAgent, when enable_isolated_metadata_proxy() spawns
a metadata proxy agent for a network, it will pass router_id
instead of network_id if metadata network is enabled and
a router port is connected to this network.

Later, MetadataDriver will register this uuid (i.e. router_id)
with monitor for the new metadata proxy process.

But when disable_isolated_metadata_proxy() destroys a metadata
proxy agent for a network, it always passes network_id as the uuid.
Thus MetadataDriver can not find the matching process. So the
corresponding metadata proxy agent can not be destroyed.

Closes-bug: #1526084
Change-Id: Icab94f4e09e95fa6729c458522a3fe50e850a4be