Updated openstack/openstack Project: openstack/trove  2349e192d778e1bb1ac1a260d145f074339872f6
Fixed 'client connection lost' bug in test_helper

When running tests that add data to datastores (such as replication)
an error can occur if the database is restarted during the test
(as is the case with MySQL).  Since the client is cached, this
would cause the connection to be closed, resulting in a failure.

Since creating the client connection is very quick and we don't need
to do it often, the solution was to not cache the client and just
create a new connection each time.

An error message was also updated to make it more clear if a
datastore-specific helper class used to load data couldn't be
imported correctly.

Change-Id: I90daa9efdeadecfee5231ead0f8b19bb0cbf560d