Updated openstack/openstack Project: openstack/tempest  7701ff4c54e77b68117ea65499a115d37cf183cc
Add external gateway to dvr before scheduling with dvr_snat agent

Before scheduling a dvr router with dvr_snat agent, we need to
add external gateway to the router. This would prevent some race
conditions that we are seeing with respect to the dhcp port's
availability in that dvr_snat agent host.

Closes-Bug: #1527372
Change-Id: I40b8684f6ec9ddd31753f7bbbdb364d1c0ec838a