Updated openstack/openstack Project: openstack/neutron  c62e69c6c48b9ff442e64cf188e5fd9965072a85
Remove check on dhcp enabled subnets while scheduling dvr

In check_ports_exist_on_l3agent we have an optimization fix
that checks for the subnets associated with the router and if
the subnets have dhcp enabled we go ahead and create the
router if it is a dvr_snat agent.

This was introduced in liberty since we saw some race condition
in the gate with single node failures.
It may not be completely right, since the dhcp agents can
run on non dvr_snat nodes as well.

Based on recommendation from the reviews, and a recent upstream
patch that sends notification on port create, we would want to
remove this and monitor the situation.

This would reduce the load on check_ports_exist_on_l3agent for
non dvr_snat nodes.

Depends-On: I40b8684f6ec9ddd31753f7bbbdb364d1c0ec838a
Related-Bug: #1513678

Change-Id: I0f50dc1101b2013caf03a64a4f48e2d03ea87b26