Updated openstack/openstack Project: openstack-infra/system-config  c3aa7860adf8226d18a91b6d24caced19394f90d
Allow openstack_citest 'grant option' privilege

The old version of the puppet mysql module defined the "all" privilege
as all the columns in the *_priv rows in the "db" table of the "mysql"
database, which included "Grant_priv". The new version simply uses the
normal MySQL API, which defines the "all" privilege as "Grant all
privileges at specified access level except GRANT OPTION".

In order to get back the old behavior of grant permissions on single use
slaves we use the options parameter with ['GRANT'].

Change-Id: I93b9375de916ca701399aee03dfaa9ab4d197319