Updated openstack/openstack Project: openstack/ceilometer  bdb352c550c8d856a255820bf32c563a620a3542
Changes aggregator transformer to allow retention_time w/o size

When using an aggregator transformer in pipeline.yaml that
looks like this:
---
transformers:
    - name: "aggregator"
      parameters:
          retention_time: 60
          resource_metadata: last
---
'size' will automatically be defaulted to 1. That means that
aggregation will not happen as expected with a retention_time of
60 seconds.

This fix will only default sample size to 1 if neither
retention_time or size are defined.

Change-Id: I4a3aa0f6de26173e6f9383d570ff2cf13d367e38
Closes-Bug: 1531626