Updated openstack/openstack Project: openstack/nova  d507474e3ac4452dafd4808efce7a78f1c3f582f
Add caching of service_min_versions in the conductor

Since we may need this information in the conductor a lot, as it has to
be used to determine whether data migration can be done, we cache the
results.

The cache gets cleared on SIGHUP, so any update of a service would
require a SIGHUP to the conductors Now (similar to how it's done for the
compute task service).

We enable this caching only for the conductor nodes as it is not needed
elsewhere.

Change-Id: I79e2f0e79e17141052e7a8128e0d64975e0b4b5f