Updated openstack/openstack Project: openstack/neutron  9ebc4f0453a4cfcdb00e728a5664769d48506a47
make floating IP specification test robust to races

The get_unused_ip function returns an address that isn't in use on
the network at the time it's called. However, by the time the caller
uses it, it may have been allocated to another concurrent test if
the network is shared or external.

The test_create_floatingip_with_specified_ip_address test creates
a floating IP with an unused IP on the external network, which has
lots of activity on it from other tests so it would occasionally
have its address taken by another test.

This patch just catches the in use error and tries again with a
new IP.

Change-Id: I5a2f78a6da1e400dea07949f35964abb767edbb2
Closes-Bug: #1532033