Updated openstack/openstack Project: openstack/horizon  8ffbd6e33be320ff28c425c9a127dcf169d656b4
Set target to _self for Header Links

Angular Panels have enabled ngRoute and HTML5 which
allows Angular to handle routing which makes all
routes to pass via Angular.

Links that do not have route defintion as a result stop
working.
This can be fixed by specifying the target="_self" on the <a> tags.

This is already in place for the Sidebar. It needs to done for the
Header Links as well.

Closes-Bug: #1531734
Change-Id: Ida76b40fdf0b66ba3ee3af4d32701974c312af57