Updated openstack/openstack Project: openstack/keystonemiddleware  66f31474700a8ec9ec383af364662e840acf8335
Adding parse of protocol v4 of AWS auth to ec2_token

This patch adds parsing of protocol v4 of AWS auth to ec2_token.
This code 'copy-pasted' from nova where it works well.
Also this patch adds unit tests for ec2_token middleware.

Chunks of the code proposed can be seen here:
https://github.com/openstack/ec2-api/blob/master/ec2api/api/__init__.py#L166

and here:
https://github.com/openstack/ec2-api/blob/master/ec2api/api/faults.py

We copy and paste the code since pulling in ec2api would bring in
a lot of dependencies and probably create a circular one

https://github.com/openstack/ec2-api/blob/master/requirements.txt

Change-Id: Id03a7c78152bda35879550f2aaf94483b82f381e
Closes-Bug: 1473039
Closes-Bug: 1333951