Updated openstack/openstack Project: openstack/neutron  ebb0dffd2f577ed100395bf89aa1702f4b1ddec3
Rename _get_vm_port_hostid in dvr to reflect the right functionality

_get_vm_port_hostid returns the host id associated with the
dvr service ports. It not only returns the VM port hostid, but
also returns the host id associated with LBaaS vip port and
dhcp.

So to be in sync with other function names let us rename the
function to _get_dvr_service_port_hostid.

Change-Id: Idc69d3b35a371b987aa34597ed6e1de7b96ff2e5