Updated openstack/openstack Project: openstack/nova-specs  c87e650c3c80b7f7e8f53292601ddc84e828955b
Corrects the API propose for live-migration progress

Same as the propose of blueprint abort-live-migration and blueprint
pause-vm-during-live-migration, the migration resource should be a
sub-resource of server '/servers/{id}/migrations'. The top-level
resource 'os-migrations' won't be extended anymore, and may totally
replaced by sub-resource migrations of server in the future.

This patch corrects the spec to add new fields into the
'/servers/{id}/migrations'. And clear the json-schema of new fields.
Also add new novaclient to the propose.

Also proposes adding the link to the old top-level resource
'/os-migrations', which also required add GET method for each migration
sub-resource of servers.

Partiail blueprint live-migration-progress-report

Change-Id: I6c2ed639cf8184466277c637d9e63982dde9ad84